Implemented Fuel modification

+++ SRVehicleDesigner/DAL/Chassis.cs
@@ -56,9 +56,9 @@ namespace SRVehicleDesigner
         [DataMember(Order = 5)]
         public int Armor { get; private set; }
         [DataMember(Order = 6)]
-        public decimal CargoFactorBase { get; private set; }
+        public int CargoFactorBase { get; private set; }
         [DataMember(Order = 7)]
-        public decimal CargoFactorMax { get; private set; }
+        public int CargoFactorMax { get; private set; }
         [DataMember(Order = 8)]
         public int AutoNav { get; private set; }
         [DataMember(Order = 9)]
